SERVER/requirements.txt
asgiref==3.5.2
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
gcloud==0.18.3
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
jws==0.1.3
mypy-extensions==0.4.3
mysqlclient==2.1.1
oauth2client==4.1.3
pathspec==0.9.0
platformdirs==2.5.2
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.15.0
PyJWT==2.4.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-jwt==2.0.1
pytz==2022.2.1
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
twilio==7.12.1
urllib3==1.26.11
whitenoise==6.2.0
